Merge branch 'backend' into 'develop'

See merge request s08-webmobile1-sub2/S08P12A804!117

+++ backend/src/main/java/com/ssafy/kkini/repository/MemoryRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MemoryRepository extends JpaRepository<Memory, Long> {
+public interface MemoryRepository extends JpaRepository<Memory, Integer> {
     List<Memory> findByUser_UserId(int userId);
 
     Optional<Memory> findByMemoryId(int memoryId);
